+++ include/cutlass/epilogue/thread/linear_combination_relu.h
@@ -18,7 +18,7 @@
  * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
  * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
  * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
- * STRICT LIABILITY, OR TOR (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
  **************************************************************************************************/
@@ -383,15 +383,7 @@ public:
     // Compute threshold optionally
     intermediate = relu(threshold_, intermediate);
 
-    if (platform::is_same<ElementOutput, int32_t>::value ||
-        platform::is_same<ElementOutput, uint32_t>::value ||
-        platform::is_same<ElementOutput, int16_t>::value ||
-        platform::is_same<ElementOutput, uint16_t>::value ||
-        platform::is_same<ElementOutput, int8_t>::value ||
-        platform::is_same<ElementOutput, uint8_t>::value ||
-        platform::is_same<ElementOutput, cutlass::int4b_t>::value ||
-        platform::is_same<ElementOutput, cutlass::uint4b_t>::value ||
-        platform::is_same<ElementOutput, cutlass::uint1b_t>::value) {
+    if (platform::numeric_limits<ElementOutput>::is_integer) {
       // Convert floats back to INT
       FragmentAccumulator scaled_accumulator;
 
@@ -432,15 +424,7 @@ public:
     // Compute threshold optionally
     intermediate = relu(threshold_, intermediate);
 
-    if (platform::is_same<ElementOutput, int32_t>::value ||
-        platform::is_same<ElementOutput, uint32_t>::value ||
-        platform::is_same<ElementOutput, int16_t>::value ||
-        platform::is_same<ElementOutput, uint16_t>::value ||
-        platform::is_same<ElementOutput, int8_t>::value ||
-        platform::is_same<ElementOutput, uint8_t>::value ||
-        platform::is_same<ElementOutput, cutlass::int4b_t>::value ||
-        platform::is_same<ElementOutput, cutlass::uint4b_t>::value ||
-        platform::is_same<ElementOutput, cutlass::uint1b_t>::value) {
+    if (platform::numeric_limits<ElementOutput>::is_integer) {
       // Convert floats back to INT
       FragmentAccumulator scaled_accumulator;